move vmdk generator out of index.js
